-changed authority assignment logic to allow all machine in system (i.e., server AND clients alike) to assign their own GONetId values when they are the owner instead of waiting on the server to assign like before....this change now makes it true that no matter what...by the time other MonoBehaviour instances installed on a GameObject with GNP on it, that when their Start() method runs, they know the values for GONetId, OwnerAuthorityId and WasInstantiated are finalized! -changed data type of OwnerAuthorityId from uint to ushort since we now limit value range down to only 10 bits (which, with logic and all....one server can support 1,022 clients)

+++ Assets/GONet/Code/GONet/GONetConnections.cs
@@ -33,7 +33,7 @@ namespace GONet
 
         static readonly ConcurrentDictionary<Thread, ArrayPool<byte>> messageByteArrayPoolByThreadMap = new ConcurrentDictionary<Thread, ArrayPool<byte>>();
 
-        public uint OwnerAuthorityId { get; internal set; }
+        public ushort OwnerAuthorityId { get; internal set; }
 
         #region round trip time stuffs (RTT)
